Logistic Discriminant Metric Learning (LDML).

Distance Metric Learning through the likelihood maximization of a logistic based probability distribution.

LDML(num_dims = NULL, b = 0.001, learning_rate = "adaptive",
  eta0 = 0.3, initial_metric = NULL, max_iter = 10, prec = 0.001,
  tol = 0.001, descent_method = "SDP", eta_thres = 1e-14,
  learn_inc = 1.01, learn_dec = 0.5)

Arguments

num_dims
Number of dimensions for dimensionality reduction. Not supported yet.

b
Logistic function positive threshold.

learning_rate
Type of learning rate update for gradient descent. Possible values are: - 'adaptive' : the learning rate will increase if the gradient step is succesful, else it will decrease. - 'constant' : the learning rate will be constant during all the gradient steps.

eta0
The initial value for learning rate. Float.

initial_metric
If array or matrix, it must be a positive semidefinite matrix with the starting metric for gradient descent, where d is the number of features. If None, euclidean distance will be used. If a string, the following values are allowed: - 'euclidean' : the euclidean distance. - 'scale' : a diagonal matrix that normalizes each attribute according to its range will be used.

max_iter
Maximum number of iterations of gradient descent. Integer.

prec
Precision stop criterion (gradient norm). Float.

tol
Tolerance stop criterion (difference between two iterations). Float.

descent_method
The descent method to use. Allowed values are: - 'SDP' : semidefinite programming, consisting of gradient descent with projections onto the PSD cone.

eta_thres
A learning rate threshold stop criterion.

learn_inc
Increase factor for learning rate. Ignored if learning_rate is not 'adaptive'.

learn_dec
Decrease factor for learning rate. Ignored if learning_rate is not 'adaptive'.

Value

The MCML transformer, structured as a named list.

References

Matthieu Guillaumin, Jakob Verbeek and Cordelia Schmid. “Is that you? Metric learning approaches for face identification”. In: Computer Vision, 2009 IEEE 12th international conference on. IEEE. 2009, pages 498-505.