dataType R6 class.
Source: R/RDML.types.R

Inherits: rdmlBaseType.

dataType

Format

An R6Class generator object.

Initialization

dataType$new(tar, cq = NULL, excl = NULL,
adp = NULL, mdp = NULL, endPt = NULL, bgFluor = NULL, bgFluorSlp = NULL,
quantFluor = NULL)

Fields

tar

idReferencesType. TargetID - A reference to a target.

cq

checkNumber. Calculated fractional PCR cycle used for downstream quantification. Negative values express following condition: Not Available: -1.0

excl

checkString. Excluded. If excl is present, this entry should not be evaluated. Do not set this element to FALSE if the entry is valid. Instead, leave the entire excl element out instead. It may contain a string with a reason for the exclusion. Several reasons for exclusion should be seperated by semicolons ";".

adp

adpsType.

mdp

mdpsType.

endPt

checkNumber

. Value of the endpoint measurement.
bgFluor

checkNumber. Background fluorescence (the y-intercept of the baseline trend based on the estimated background fluorescence).

bgFluorSlp

checkNumber. Background fluorescence slope - The slope of the baseline trend based on the estimated background fluorescence. The element should be absent to indicate a slope of 0.0; If this element is present without the bgFluor element it should be ignored.

quantFluor

checkNumber. Quantification flourescence - The fluorescence value corresponding to the treshold line.

Methods

AsDataFrame(dp.type = "adp")

Represents amplification (
dp.type = "adp"
) or melting (dp.type = "mdp") data points as data.frame