runType R6 class.

A run is a set of reactions performed in one "run", for example one plate, one rotor, one array, one chip. Inherits: rdmlBaseType.

runType

Format

An R6Class generator object.

Initialization

runType$new(id, description = NULL,
  documentation = NULL, experimenter = NULL, instrument = NULL,
  dataCollectionSoftware = NULL, backgroundDeterminationMethod = NULL,
  cqDetectionMethod = NULL, thermalCyclingConditions = NULL, pcrFormat,
  runDate = NULL, react = NULL)

Fields

id

idType.

description

checkString.

documentation

list of idReferencesType.

experimenter

list of idReferencesType.

instrument

checkString. Description of the instrument used to aquire the data.

dataCollectionSoftware

dataCollectionSoftwareType. Description of the software used to analyze/collect the data.

backgroundDeterminationMethod

checkString. Description of method used to determine the background.

cqDetectionMethod

cqDetectionMethodType. Description of method used to calculate the quantification cycle.

thermalCyclingConditions

idReferencesType. The program used to aquire the data.

pcrFormat

adpsType.

runDate

adpsType. Time stamp of data acquisition.

react

list of adpsType.

Methods

AsDataFrame(dp.type = "adp")

Represents amplification (dp.type = "adp") or melting (dp.type = "mdp") data points as data.frame