adpsType R6 class.

adpsType R6 class.

adpsType

Format

An R6Class generator object.

Details

Contains matrix of amplification data. Must have three columns:

cyc

PCR cycle at which data point was collected (every cycle must have unique number).

tmp

temperature in degrees Celsius at the time of measurement (optional).

fluor

raw fluorescence intensity measured.

Inherits: rdmlBaseType.

Initialization

adpsType$new(fpoints)

Fields

fpoints

assertMatrix. Matrix with amplification data points.

Examples

#cycles
cyc <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17,
18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33,
34, 35, 36, 37, 38, 39, 40)
#fluorescence
fluo <- c(2.0172, 2.0131, 2.0035, 2, 2.0024, 2.0056, 2.0105, 2.0179,
2.0272, 2.0488, 2.0922, 2.1925, 2.3937, 2.7499, 3.3072, 4.0966,
5.0637, 6.0621, 7.0239, 7.8457, 8.5449, 9.1282, 9.6022, 9.9995,
10.2657, 10.4989, 10.6813, 10.8209, 10.9158, 10.9668, 11.0053,
11.0318, 11.0446, 11.044, 11.0052, 10.9671, 10.9365, 10.9199,
10.897, 10.8316)
#temperature
temp <- c(55, 55, 55, 55, 54, 54, 55, 55, 55, 55, 55, 55, 55, 55, 55,
55, 55, 55, 55, 55, 55, 55, 55, 56, 55, 55, 55, 55, 55, 55, 55,
55, 55, 55, 55, 55, 55, 55, 55, 55)

#combine all variables into a proper object
data <- data.frame(cyc = cyc, tmp = temp, fluor = fluo)

#create adps object
adpsType$new(data)

#> 	fpoints: [cyc, tmp, fluor]

#create adps object without temperature data
adpsType$new(data[, -2])

#> 	fpoints: [cyc, fluor]