Creates new instance of RDML class object

This function has been designed to import data from RDML v1.1 and v1.2 format files or from xls file generated by Applied Biosystems 7500. To import from xls this file have to contain Sample Setup and Multicomponent Data sheets!

Arguments

filename
string -- path to file

show.progress
logical -- show loading progress bar if TRUE

conditions.sep
separator for condition defined at sample name

format
string -- input file format. Possible values auto, rdml, abi, excel, csv. See Details.

Details

File format options:

abi

Reads .xls files generated by ABI 7500 v.2. To create such files use File>Export; check 'Sample Setup' and 'Multicomponent Data'; select 'One File'

excel

.xls or .xslx file with sheets 'description', 'adp', 'mdp'. See example file table.xlsx

csv

.csv file with first column 'cyc' or 'tmp' and fluorescence data in other columns

rdml

.rdml or .lc96p files

Warning

Although the format RDML claimed as data exchange format, the specific implementation of the format at devices from real manufacturers differ significantly. Currently this function is checked against RDML data from devices: Bio-Rad CFX96, Roche LightCycler 96 and Applied Biosystems StepOne.

Examples

not_run({
  ## Import from RDML file
  PATH <- path.package("RDML")
  filename <- paste(PATH, "/extdata/", "lc96_bACTXY.rdml", sep ="")
  lc96 <- RDML$new(filename)

  ## Some kind of overview for lc96
  lc96$AsTable(name.pattern = sample[[react$sample$id]]$description)
  lc96$AsDendrogram()
})