pcrFormatType R6 class.
Source: R/RDML.types.R

The display format of the PCR, analogous to the the qPCR instrument run format. Inherits: rdmlBaseType.

pcrFormatType

Format

An R6Class generator object.

Details

Rotor formats always have 1 column; rows correspond to the number of places in the rotor. Values for common formats are:

Format 	 rows 	 columns 	 rowLabel 	 columnLabel

single-well 	 1 	 1 	 123 	 123

48-well plate 	 6 	 8 	 ABC 	 123

96-well plate 	 8 	 12 	 ABC 	 123

384-well plate 	 16 	 24 	 ABC 	 123

1536-well plate 	 32 	 48 	 ABC 	 123

3072-well array 	 32 	 96 	 A1a1 	 A1a1

5184-well chip 	 72 	 72 	 ABC 	 123

32-well rotor 	 32 	 1 	 123 	 123

72-well rotor 	 72 	 1 	 123 	 123

100-well rotor 	 100 	 1 	 123 	 123

If rows field has value -1, the function will not try to reconstruct a plate and just display all run data in a single column. If the columns field has value 1 then the function will not display a column label.

Initialization

pcrFormatType$new(rows, columns, rowLabel, columnLabel)

@section Fields:

rows

checkCount.

columns

checkCount.

rowLabel

labelFormatType.

columnLabel

labelFormatType.