reactType R6 class.

A reaction is an independent chemical reaction corresponding for example to a well in a 96 well plate, a capillary in a rotor, a through-hole on an array, etc. Inherits: rdmlBaseType.

reactType

Format

An R6Class generator object.

Details

The ID of this reaction

Schemas :

rotor : assign IDs according to the position of the sample on the rotor (1 for the 1st sample, 2 for the 2nd, ...)

plate (96/384/1536 well) : the IDs are assigned in a row-first/column-second manner. For each row, the samples are numbered according to the increasing column number. At the end of a row, the numbering starts at the first column of the next row. An example for this type of plate can be found below :

1 	 2 	 3 	 ...

A 	 1 	 2 	 3

B 	 13 	 14

or

1 	 2 	 3 	 ...

1 	 1 	 2 	 3

2 	 13 	 14

multi-array plate (BioTrove) : the IDs are assigned in a row-first/column-second manner, ignoring the organisation of sub-arrays. For each row, the samples are numbered according to the increasing column number. At the end of a row, the the next row. An example for this type of plate can be found below : todo...

Initialization

reactType$new(id, sample, data = NULL, pcrFormat = pcrFormatType$new(8, 12, labelFormatType$new("123"), labelFormatType$new("ABC")))

@section Fields:

id

reactIdType. See 'Details'.

sample

idReferencesType. SampleID - A reference to a sample.

data

list of dataType.

position

Human readable form of the react id (i.e. '13' -> 'B1')..

Methods

.recalcPosition(pcrformat)

Converts react id to the human readable form (i.e. '13' -> 'B1'). This converted value can be accessed by position field. pcrFormat is pcrFormatType. Currently, only 'ABC' and '123' are supported as labels. For '123' '123' the Position will look like 'r01c01', for 'ABC' '123' it will be 'A01' and for '123' 'ABC' it will be 01A. 'ABC' 'ABC' is not currently supported. Note that 'ABC' will result in loss of information if the experiment contains more than 26 rows!